Create slack.js
This code is responsible for sending a message to a **Slack channel** using a **Slack Incoming Webhook**. It imports the `axios` library to make HTTP requests. The function `sendToSlack` takes a `message` as input and sends it as a POST request to the Slack webhook URL, which is stored securely in the environment variables (`SLACK_WEBHOOK_URL`). The message is sent in plain text format to the configured Slack channel. This function is exported so it can be used in other parts of the application, such as when a summary of to-dos is generated and needs to be shared with a team via Slack.