Create api.js
This code sets up a reusable **Axios instance** for making HTTP requests to the backend server. It sets the base URL to `http://localhost:5000`, which means all API calls made using this instance will automatically use this base path. This helps simplify the code in other files, as you don’t need to repeat the full server URL every time you make a request. It's used to connect the frontend with the backend API.